Is returning a SOP (without arbitrary).
Arbitrary relation is removed and parallel is used instead, to make it simpler for others to parse. Changed in RelateTwoOperations.java line 61-62!

Limited to only run one operation at the time since no childs or arbitrary are used. All operations in booking a shared variable. This is created in CreateBooking.java. This is not applied to the planner!

Fixed bug in compleGetComplParallelOps.java which didn't work when only variables are used.

Fixed bug in the SOP xml-generation.

Added ConditionToJavaStringParser.java that creates java-style conditions for the planner. Also fixed a bugg in the toString in ConditionExpression.java